moved run_kopf to be shared by all
